CI: Test with Atom 1.56 built against Electron 8
With Atom at DeeDeeG's custom commit
0a668c8893da8c24324c6969de99fc0c38fbbde2.

Built against Electron 8.5.5.

Built with the CI base image's included Visual Studio Enterprise 2017,
as the CLI installer for Visual Studio 2015 has stopped working (???).

Built with apm 2.6.2, so as to allow Atom to bootstrap/build in CI
with Visual Studio 2017 instead of trying to use the busted included
copy of Visual Studio 2015 on the CI base image.

Based closely on the upstream Atom v1.56.0 tag,
with the above-mentioned modifications.